# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Incident Management",
        "description": "Incident Management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api",
            "description": "local server"
        },
        {
            "url": "#",
            "description": "staging server"
        },
        {
            "url": "#",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/incidents": {
            "get": {
                "tags": [
                    "Incident"
                ],
                "summary": "Get all incidents",
                "description": "Retrieve a paginated list of all incidents",
                "operationId": "c58b1625d41eb05fa2809fa6eb326e25",
                "responses": {
                    "200": {
                        "description": "A list of incidents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "incidents": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 35
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Exemplo swagger"
                                                            },
                                                            "evidence": {
                                                                "type": "string",
                                                                "example": "Nosso e-mail corporativo chega pishing a todo momento, precisamos de ajuda para bloquear alguns remetentes"
                                                            },
                                                            "criticality": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "host": {
                                                                "type": "string",
                                                                "example": "ambev.tech.br"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-09-05T17:41:58.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-09-05T17:41:58.000000Z"
                                                            },
                                                            "user_id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "deleted_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "null"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "first_page_url": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1/api/incidents?page=1"
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page_url": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1/api/incidents?page=1"
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "example": "null"
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "&laquo; Previous"
                                                            },
                                                            "active": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1/api/incidents"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Incident"
                ],
                "summary": "Create a new incident",
                "description": "Submit a new incident for processing",
                "operationId": "9c3c4a3c596429e52f3580bcae9c8a2a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "evidence",
                                    "criticality",
                                    "host",
                                    "user_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Exemplo swagger"
                                    },
                                    "evidence": {
                                        "type": "string",
                                        "example": "Nosso e-mail corporativo chega pishing a todo momento, precisamos de ajuda para bloquear alguns remetentes"
                                    },
                                    "criticality": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "host": {
                                        "type": "string",
                                        "example": "ambev.tech.br"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Incident being processed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Incidente sendo processado!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Incidente não cadastrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}